requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.7
langchain-openai>=0.0.5
langchain-chroma>=0.0.6

# Vector store
chromadb>=0.4.22

# OpenAI
openai>=1.12.0

# Text processing
nltk>=3.8.1

# Environment variables
python-dotenv>=1.0.0

# Type hints support
typing-extensions>=4.9.0

# Optional: for better handling of file paths
pathlib>=1.0.1

# New requirements for embedding comparison
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0